feat: package structure
Update structure to match other Boundary plugins which groups packages by service
